+++ src/styles/colorMap.ts
@@ -1,0 +1,12 @@
+export const colorMap = {
+  blue: "#6698F5",
+  ocean: "#5AA6C7",
+  lavender: "#949AEC",
+  mint: "#77CEC6",
+  sage: "#AECA99",
+  gray: "#A9A9A9",
+  orange: "#FDB456",
+  coral: "#FD855F",
+  rose: "#ED83B1",
+  plum: "#D49AE9",
+};
